src/requirements.txt
torch>=1.8.0
lap>=0.4.0
scikit-learn>=1.1.3
numpydoc>=1.1
anndata>=0.8.0
scanpy>=1.9.1
scipy>=1.9.3
numba>=0.14.0
tqdm
